# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * RomeGuest AI - Professional Host Suite API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  slug: string;
  name: string;
  content: string;
  /** @nullable */
  whatsappNumber: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePropertyRequest {
  /** Unique URL-friendly identifier (e.g. 'fleming-1') */
  slug: string;
  /** Human-readable property name */
  name: string;
  /** House rules, WiFi info, tips, etc. */
  content: string;
  /** Host WhatsApp number for SOS button */
  whatsappNumber?: string;
}

export interface UpdatePropertyRequest {
  /** Property-specific host password */
  hostPassword?: string;
  name?: string;
  content?: string;
  whatsappNumber?: string;
}

export type ConversationMessageRole =
  (typeof ConversationMessageRole)[keyof typeof ConversationMessageRole];

export const ConversationMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ConversationMessage {
  role: ConversationMessageRole;
  content: string;
}

export interface ChatMessageRequest {
  message: string;
  conversationHistory?: ConversationMessage[];
}

export interface ChatMessageResponse {
  reply: string;
  propertyName: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * HMAC session token from POST /auth/ceo-login
 */
export type CeoSessionHeaderParameter = string;