# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: RomeGuest AI - Professional Host Suite API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: properties
    description: Property management
  - name: chat
    description: Guest chat operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties (CEO only)
      parameters:
        - $ref: "#/components/parameters/CeoSessionHeader"
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property (CEO only)
      parameters:
        - $ref: "#/components/parameters/CeoSessionHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyRequest"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /properties/{slug}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Property not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProperty
      tags: [properties]
      summary: Update property content
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CeoSessionHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyRequest"
      responses:
        "200":
          description: Property updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property (CEO only)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CeoSessionHeader"
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /properties/{slug}/chat:
    post:
      operationId: sendPropertyChat
      tags: [chat]
      summary: Send a chat message for a specific property
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageResponse"
        "404":
          description: Property not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    CeoSessionHeader:
      name: X-CEO-Session
      in: header
      required: true
      description: HMAC session token from POST /auth/ceo-login
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Property:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        content:
          type: string
        whatsappNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - content
        - whatsappNumber
        - createdAt
        - updatedAt

    CreatePropertyRequest:
      type: object
      properties:
        slug:
          type: string
          description: Unique URL-friendly identifier (e.g. 'fleming-1')
        name:
          type: string
          description: Human-readable property name
        content:
          type: string
          description: House rules, WiFi info, tips, etc.
        whatsappNumber:
          type: string
          description: Host WhatsApp number for SOS button
      required:
        - slug
        - name
        - content

    UpdatePropertyRequest:
      type: object
      properties:
        hostPassword:
          type: string
          description: Property-specific host password
        name:
          type: string
        content:
          type: string
        whatsappNumber:
          type: string
      required: []

    ChatMessageRequest:
      type: object
      properties:
        message:
          type: string
        conversationHistory:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMessage"
      required:
        - message

    ChatMessageResponse:
      type: object
      properties:
        reply:
          type: string
        propertyName:
          type: string
      required:
        - reply
        - propertyName

    ConversationMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error