# lib/api-zod/src/generated/types/chatMessageRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * RomeGuest AI - Professional Host Suite API
 * OpenAPI spec version: 0.2.0
 */
import type { ConversationMessage } from "./conversationMessage";

export interface ChatMessageRequest {
  message: string;
  conversationHistory?: ConversationMessage[];
}