# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * RomeGuest AI - Professional Host Suite API
 * OpenAPI spec version: 0.2.0
 */

export interface Property {
  id: number;
  slug: string;
  name: string;
  content: string;
  /** @nullable */
  whatsappNumber: string | null;
  createdAt: Date;
  updatedAt: Date;
}