# lib/api-zod/src/generated/types/updatePropertyRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * RomeGuest AI - Professional Host Suite API
 * OpenAPI spec version: 0.2.0
 */

export interface UpdatePropertyRequest {
  /** Property-specific host password */
  hostPassword?: string;
  name?: string;
  content?: string;
  whatsappNumber?: string;
}